requirements.txt
# =========================
# CORE
# =========================
pandas==2.1.3
matplotlib==3.8.2
numpy==1.26.4

# =========================
# STREAMLIT
# =========================
streamlit==1.54.0
streamlit-calendar==0.7.0
plotly==5.18.0

# =========================
# OPENAI + LANGCHAIN
# =========================
openai==2.20.0
langchain==1.2.10
langchain-community==0.4.1
langchain-openai==1.1.9

# =========================
# UTILIDADES
# =========================
python-dotenv==1.2.1
python-decouple==3.8

# =========================
# GMAIL API
# =========================
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0